+++ test/fixtures/cjs-3rdparty.cjs
@@ -1,0 +1,6 @@
+const resolve = require('resolve');
+
+require('./cjs2.cjs');
+
+// eslint-disable-next-line no-console
+console.log('resolve', resolve);